Temp allocator

+++ example/types.d.ts
@@ -1,7 +1,7 @@
 import * as wasm from "../wasm/runtime.js"
 
 export interface WasmExports extends wasm.OdinExports {
-	frame: (delta: wasm.i32, ctx_ptr: wasm.rawptr) => void
+	frame: (delta: wasm.f32, ctx_ptr: wasm.rawptr) => void
 	on_window_resize: (
 		window_w: wasm.f32,
 		window_h: wasm.f32,
